pyproject.toml
[tool.poetry]
name = "greenbot"
version = "0.1.0"
description = ""
authors = ["Sapphire <artevenia@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.3.2"
aiolimiter = "^1.1.0"
asyncpg = "^0.29.0"
lxml = "^5.2.2"
aiofiles = "^23.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
